Funcionalidad de alta baja y modificacion de cuestionario
Funcionalidad de alta baja y modificacion de cuestionario

+++ app/cuestionario/CuestionarioController.php
@@ -25,63 +25,76 @@ class CuestionarioController extends BaseController
     }
     private function almacenarCuestionario($params)
     {
-        if ($this->getRequestMethod() == "POST") {
-            // if ($this->getPerfilUsuario() == 3) {
-            $tipoPreguntaService = new CuestionarioService;
-            $resultSet = $tipoPreguntaService->insertCuestionario($params);
+        if ($this->getPerfilUsuario() == 3) {
+            if ($this->getRequestMethod() == "POST") {
+                $objService = new CuestionarioService;
+                $resultSet = $objService->insertCuestionario($params);
 
-            // $resultSet = $this->obtenerDatosTabla($params["idTramite"], "autorizacion");
-            if ($resultSet != null) {
-                $response = crearRespuestaSolicitud(200, "OK", "El Cuestionario se registro correctamente.", $resultSet);
+                // $resultSet = $this->obtenerDatosTabla($params["idTramite"], "autorizacion");
+                if ($resultSet != null) {
+                    $response = crearRespuestaSolicitud(200, "OK", "El Cuestionario se registro correctamente.", $resultSet);
+                } else {
+                    $response = crearRespuestaSolicitud(400, "error", "No se pudo registrar el cuestionario.");
+                }
+                $response['headers'] = ['HTTP/1.1 200 OK'];
             } else {
-                $response = crearRespuestaSolicitud(400, "error", "No se pudo registrar el cuestionario.");
+                $response = crearRespuestaSolicitud(400, "error", "Metodo HTTP equivocado.");
             }
-            $response['headers'] = ['HTTP/1.1 200 OK'];
-            // } else {
-            //     $response = crearRespuestaSolicitud(401, "error", "No tiene permisos para ejecutar la consulta.");
-            //     $response['headers'] = ['HTTP/1.1 401 Unauthorized'];
-            // }
         } else {
-            $response = crearRespuestaSolicitud(400, "error", "Metodo HTTP equivocado.");
+            $response = crearRespuestaSolicitud(401, "error", "No tiene permisos para ejecutar la consulta.");
+            $response['headers'] = ['HTTP/1.1 401 Unauthorized'];
         }
         return $response;
     }
     private function actualizarCuestionario($params)
     {
-        if ($this->getRequestMethod() == "PUT") {
-
-            //obtenemos el nombre del campo donde vamos a registrar el valor
-            $objService = new CuestionarioService();
-            $affectedRows = $objService->updateCuestionario($params);
-            if ($affectedRows > 0) {
-                $response = crearRespuestaSolicitud(200, "OK", "Se modifico el cuestionario.", ["affectedRows" => $affectedRows]);
+        if ($this->getPerfilUsuario() == 3) {
+            if ($this->getRequestMethod() == "PUT") {
+                if (isset($params['id_cuestionario'])) {
+                    //obtenemos el nombre del campo donde vamos a registrar el valor
+                    $objService = new CuestionarioService();
+                    $affectedRows = $objService->updateCuestionario($params);
+                    if ($affectedRows > 0) {
+                        $response = crearRespuestaSolicitud(200, "OK", "Se modifico el cuestionario.", ["affectedRows" => $affectedRows]);
+                    } else {
+                        $response = crearRespuestaSolicitud(400, "error", "Ocurrió un error al intentar modificar el cuestionario.");
+                    }
+                } else {
+                    $response = crearRespuestaSolicitud(400, "error", "Ocurrió un error al intentar eliminar el cuestionario.");
+                }
             } else {
-                $response = crearRespuestaSolicitud(400, "error", "Ocurrió un error al intentar modificar el cuestionario.");
+                //$response['headers'] = ['HTTP/1.1 400 Bad Request'];
+                $response = crearRespuestaSolicitud(400, "error", "Metodo HTTP equivocado");
             }
         } else {
-            //$response['headers'] = ['HTTP/1.1 400 Bad Request'];
-            $response = crearRespuestaSolicitud(400, "error", "Metodo HTTP equivocado");
+            $response = crearRespuestaSolicitud(401, "error", "No tiene permisos para ejecutar la consulta.");
+            $response['headers'] = ['HTTP/1.1 401 Unauthorized'];
         }
         return $response;
     }
     private function eliminarCuestionario($params)
     {
-        if ($this->getRequestMethod() == "PUT") {
-            if (isset($params['id_cuestionario'])) {
-                //obtenemos el nombre del campo donde vamos a registrar el valor
-                $objService = new CuestionarioService();
-                $affectedRows = $objService->deleteCuestionario($params);
-                if ($affectedRows > 0) {
-                    $response = crearRespuestaSolicitud(200, "OK", "Se elimino el cuestionario.", ["affectedRows" => $affectedRows]);
+        if ($this->getPerfilUsuario() == 3) {
+            if ($this->getRequestMethod() == "PUT") {
+                if (isset($params['id_cuestionario'])) {
+                    //obtenemos el nombre del campo donde vamos a registrar el valor
+                    $objService = new CuestionarioService();
+                    $affectedRows = $objService->deleteCuestionario($params);
+                    if ($affectedRows > 0) {
+                        $response = crearRespuestaSolicitud(200, "OK", "Se elimino el cuestionario.", ["affectedRows" => $affectedRows]);
+                    } else {
+                        $response = crearRespuestaSolicitud(400, "error", "Ocurrió un error al intentar eliminar el cuestionario.");
+                    }
                 } else {
                     $response = crearRespuestaSolicitud(400, "error", "Ocurrió un error al intentar eliminar el cuestionario.");
                 }
             } else {
-                $response = crearRespuestaSolicitud(400, "error", "Ocurrió un error al intentar eliminar el cuestionario.");
+                //$response['headers'] = ['HTTP/1.1 400 Bad Request'];
+                $response = crearRespuestaSolicitud(400, "error", "Metodo HTTP equivocado");
             }
         } else {
-            //$response['headers'] = ['HTTP/1.1 400 Bad Request'];
-            $response = crearRespuestaSolicitud(400, "error", "Metodo HTTP equivocado");
+            $response = crearRespuestaSolicitud(401, "error", "No tiene permisos para ejecutar la consulta.");
+            $response['headers'] = ['HTTP/1.1 401 Unauthorized'];
         }
         return $response;
     }
